+++ site/dat/docs/Changelog.md
@@ -1,8 +1,16 @@
 # Changelog
 
-## 0.4.2 (next)
- - bump up gatling version to 2.1.7
+## 0.4.3 (next)
+ - allow passing config as stdin to CLI like `./generate-config.sh | bzt`
+ - allow having `report-name=ask` for BlazeMeter reporter
+
+## 0.4.2
+ - bump up Gatling version to 2.1.7
  - selenium script from requests format
+ - change file search algo to use paths relative to config locations
+ - allow having script files relative to config locations
+ - allow having `included-configs` instruction
+ - allow setting path to artifacts from `settings` in config, deprecate `--data-dir` CLI option
 
 ## 0.4.1
  - fix Locust.io search paths